Moved the scopes to a subpackage
