+++ src/main/java/edu/clemson/cs/rsrg/typeandpopulate/query/searchpath/ScopeSearchPath.java
@@ -15,8 +15,8 @@ package edu.clemson.cs.rsrg.typeandpopulate.query.searchpath;
 import edu.clemson.cs.rsrg.typeandpopulate.entry.SymbolTableEntry;
 import edu.clemson.cs.rsrg.typeandpopulate.exception.DuplicateSymbolException;
 import edu.clemson.cs.rsrg.typeandpopulate.query.searcher.TableSearcher;
-import edu.clemson.cs.rsrg.typeandpopulate.symboltables.Scope;
 import edu.clemson.cs.rsrg.typeandpopulate.symboltables.ScopeRepository;
+import edu.clemson.cs.rsrg.typeandpopulate.symboltables.scopes.Scope;
 import java.util.List;
 
 /**
