+++ src/main/java/edu/clemson/cs/rsrg/typeandpopulate/symboltables/ScopeRepository.java
@@ -14,6 +14,8 @@ package edu.clemson.cs.rsrg.typeandpopulate.symboltables;
 
 import edu.clemson.cs.rsrg.absyn.ResolveConceptualElement;
 import edu.clemson.cs.rsrg.typeandpopulate.exception.NoSuchSymbolException;
+import edu.clemson.cs.rsrg.typeandpopulate.symboltables.scopes.ModuleScope;
+import edu.clemson.cs.rsrg.typeandpopulate.symboltables.scopes.Scope;
 import edu.clemson.cs.rsrg.typeandpopulate.typereasoning.TypeGraph;
 import edu.clemson.cs.rsrg.typeandpopulate.utilities.ModuleIdentifier;
